// app/types/intern.ts
export type Intern = {
  intern_id: number
  title: string
  description: string
  link: string
  entry_deadline_date: string
  started_at: string
  ended_at: string
  site: string
  style_type: string
  recruitment_type: string
  hash_id: string
  capacity: string
  reward: string
  technologies: string
  necessary_requirements: string
  welcome_requirements: string
}